Moving test functions into 'commands.js'
